create package layout, add setup.py
