+++ .gitignore
@@ -1,16 +1,10 @@
-setup.py
+notebooks/
+build/
+dist/bill
+.ipynb_checkpoints/
 __pycache__
 sqlbills.egg-info
-qr-to-sql-with-modul.ipynb
-sql-explore.ipynb
-.ipynb_checkpoints
-sql-explore.py
-untitled
-notebooks
-build/
 *.db
-dist/bill
-bill.spec
-access.log
-error.log
-flask.log
+telegram_bot/.env
+telegram_bot/set-env.sh
+billdb.egg-info
