+++ api/gunicorn_config.py
@@ -1,7 +1,2 @@
-import logging
-
 bind = "0.0.0.0:5001"  # Replace with your desired host and port
 workers = 1           # You can adjust the number of workers based on your server's resources
-loglevel = "debug"
-errorlog = "./error.log"
-accesslog = "./access.log"
